llm-redteam-platform/backend/requirements.txt
# Core Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
Alembic==1.12.1

# Async Processing
Celery==5.3.4
Redis==5.0.1

# LLM APIs
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# ML/NLP Libraries
nltk==3.8.1
sentence-transformers==2.2.2
numpy==1.24.4
scikit-learn==1.3.2

# Data Processing
pandas==2.1.4
requests==2.31.0

# Security & Validation
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Development
python-socketio==5.9.0
eventlet==0.33.3